# api-docs.json
{"openapi":"3.0.1","info":{"title":"Emlakjet E-commerce Case","description":"Endpoints related to Emlakjet E-commerce Case","version":"1.0"},"servers":[{"url":"http://localhost:8080/api","description":"Generated server url"}],"security":[{"Bearer Authentication":[]}],"paths":{"/product/v1/{productId}":{"get":{"tags":["product-controller"],"operationId":"getProduct","parameters":[{"name":"productId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Product details","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductDto"}}}},"404":{"description":"Product not found","content":{"*/*":{"example":{"message":"Product with given ID doesn't exists."}}}},"401":{"description":"Authentication problem","content":{"*/*":{"example":{"message":"Authorization token is not provided or invalid"}}}}}},"put":{"tags":["product-controller"],"operationId":"updateProduct","parameters":[{"name":"productId","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductUpdateDto"}}},"required":true},"responses":{"400":{"description":"Invalid request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CommonResponse"}}}},"200":{"description":"Product updated successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductDto"}}}},"401":{"description":"Authentication problem","content":{"*/*":{"example":{"message":"Authorization token is not provided or invalid"}}}}}},"delete":{"tags":["product-controller"],"operationId":"deleteProduct","parameters":[{"name":"productId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"201":{"description":"Product deleted successfully"},"404":{"description":"Given productId does not exist","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CommonResponse"}}}},"401":{"description":"Authentication problem","content":{"*/*":{"example":{"message":"Authorization token is not provided or invalid"}}}},"204":{"description":"No Content"}}}},"/product/v1":{"get":{"tags":["product-controller"],"operationId":"listProducts","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"pageSize","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}}],"responses":{"401":{"description":"Authentication problem","content":{"*/*":{"example":{"message":"Authorization token is not provided or invalid"}}}},"200":{"description":"List of products","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PagedResponseProductDto"}}}}}},"post":{"tags":["product-controller"],"operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductDto"}}},"required":true},"responses":{"400":{"description":"Invalid request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CommonResponse"}}}},"409":{"description":"Product with given code already exists","content":{"*/*":{"example":{"message":"Another product with same code already exists."}}}},"201":{"description":"Successfully created the resource","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductDto"}}}}}}},"/invoice/v1":{"get":{"tags":["invoice-controller"],"operationId":"listInvoices","parameters":[{"name":"isApproved","in":"query","required":false,"schema":{"type":"boolean"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"pageSize","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}}],"responses":{"200":{"description":"List of invoices with given parameters","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PagedResponseInvoiceDto"}}}},"401":{"description":"Authentication problem","content":{"*/*":{"example":{"message":"Authorization token is not provided or invalid"}}}}}},"post":{"tags":["invoice-controller"],"operationId":"create_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InvoiceDto"}}},"required":true},"responses":{"201":{"description":"Resource created successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/InvoiceCreationResult"}}}},"400":{"description":"Invalid request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CommonResponse"}}}},"422":{"description":"Unprocessable entity given","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CommonResponse"}}}},"401":{"description":"Authentication problem","content":{"*/*":{"example":{"message":"Authorization token is not provided or invalid"}}}}}}},"/auth/v1/register":{"post":{"tags":["auth-controller"],"operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"201":{"description":"Successfully registered user","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}},"409":{"description":"User with given email address already exists","content":{"*/*":{"example":{"message":"User with same E-mail already exists"}}}}}}},"/auth/v1/login":{"post":{"tags":["auth-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"Successfully logged in","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}},"401":{"description":"Invalid credentials","content":{"*/*":{"example":{"message":"Invalid email or password"}}}}}}},"/invoice/v1/{invoiceId}":{"get":{"tags":["invoice-controller"],"operationId":"getInvoice","parameters":[{"name":"invoiceId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"404":{"description":"Invoice not found","content":{"*/*":{"example":{"message":"Invoice with given ID doesn't exists."}}}},"401":{"description":"Authentication problem","content":{"*/*":{"example":{"message":"Authorization token is not provided or invalid"}}}},"200":{"description":"Invoice details","content":{"*/*":{"schema":{"$ref":"#/components/schemas/InvoiceDto"}}}}}}}},"components":{"schemas":{"ProductUpdateDto":{"type":"object","properties":{"code":{"maxLength":6,"minLength":6,"type":"string","example":"P00001"},"name":{"maxLength":100,"minLength":2,"type":"string","example":"Product 1"},"description":{"maxLength":100,"minLength":2,"type":"string","example":"Very nice product"},"price":{"maximum":999999999999,"exclusiveMaximum":false,"minimum":0.01,"exclusiveMinimum":false,"type":"number","format":"double","example":50}}},"CommonResponse":{"type":"object","properties":{"message":{"type":"string","description":"Message explaining the situation","example":"User with given email address already exists"}}},"ProductDto":{"required":["code","description","name","price"],"type":"object","properties":{"id":{"type":"string","description":"ID of the product","readOnly":true,"example":"6741da3c38bac77c991e3a17"},"code":{"pattern":"\\S{6}","type":"string","description":"Six character long product code","example":"P00001"},"name":{"maxLength":100,"minLength":2,"type":"string","description":"Name of the product","example":"Product 1"},"description":{"maxLength":100,"minLength":2,"type":"string","description":"Short description of the product","example":"Very nice product"},"price":{"maximum":999999999999,"exclusiveMaximum":false,"minimum":0.01,"exclusiveMinimum":false,"type":"number","description":"Price of the product","format":"double","example":50},"createdBy":{"type":"string","description":"ID of the user who created the product","readOnly":true,"example":"6740a9f2a4cdb00fd22b6117"}}},"InvoiceDto":{"required":["amount","billNo","email","firstName","lastName","productCode"],"type":"object","properties":{"id":{"type":"string","description":"ID of the invoice","readOnly":true,"example":"6741da3c38bac77c991e3a17"},"firstName":{"maxLength":100,"minLength":2,"type":"string","description":"First name of the invoice owner","example":"John"},"lastName":{"maxLength":100,"minLength":2,"type":"string","description":"Last name of the invoice owner","example":"Doe"},"email":{"maxLength":320,"minLength":6,"type":"string","description":"Email of the invoice owner","example":"john@doe.com"},"amount":{"maximum":999999999999,"exclusiveMaximum":false,"minimum":0.01,"exclusiveMinimum":false,"type":"number","description":"Amount of the invoice","format":"double","example":100},"productCode":{"pattern":"\\S{6}","type":"string","description":"Product code of the bought item","example":"P00001"},"billNo":{"maxLength":255,"minLength":1,"pattern":"\\w{3}\\d{13}","type":"string","description":"Bill no of the invoice","example":"FYS2024000000001"},"createdBy":{"type":"string","description":"Creator of the invoice","readOnly":true,"example":"6740a9f2a4cdb00fd22b6117"},"isApproved":{"type":"boolean","description":"Is invoice approved","readOnly":true,"example":true},"product":{"$ref":"#/components/schemas/ProductDto"}}},"InvoiceCreationResult":{"type":"object","properties":{"message":{"type":"string"},"invoice":{"$ref":"#/components/schemas/InvoiceDto"},"successful":{"type":"boolean"}}},"RegisterRequest":{"required":["email","fullName","password"],"type":"object","properties":{"email":{"maxLength":320,"minLength":6,"type":"string","example":"john@doe.com"},"fullName":{"maxLength":255,"minLength":5,"type":"string","example":"John Doe"},"password":{"maxLength":255,"minLength":6,"type":"string","example":"john123"}}},"AuthResponse":{"type":"object","properties":{"token":{"type":"string","description":"JWT of the user","example":"eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJqb2huQGRvZS5jb20iLCJpYXQiOjE3MzI0NjMwNDcsImV4cCI6MTczMjQ2MzA0N30.mBWOXKlDkBykK5RrIZYQ5-5PxMIcytriPGQ2CUXiEyg"},"expiresIn":{"type":"integer","description":"Expiration time of the token in milliseconds","format":"int64","example":3600000}}},"LoginRequest":{"required":["email","password"],"type":"object","properties":{"email":{"maxLength":320,"minLength":6,"type":"string","description":"Email of the commerce user","example":"john@doe.com"},"password":{"maxLength":255,"minLength":6,"type":"string","description":"Password of the bought item","example":"john123"}}},"CustomPageable":{"type":"object","properties":{"pageNumber":{"type":"integer","format":"int32"},"pageSize":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"}}},"PagedResponseProductDto":{"type":"object","properties":{"content":{"type":"array","items":{"$ref":"#/components/schemas/ProductDto"}},"pageable":{"$ref":"#/components/schemas/CustomPageable"}}},"PagedResponseInvoiceDto":{"type":"object","properties":{"content":{"type":"array","items":{"$ref":"#/components/schemas/InvoiceDto"}},"pageable":{"$ref":"#/components/schemas/CustomPageable"}}}},"securitySchemes":{"Bearer Authentication":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}}}